% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radarWidget.R
\name{radarWidget}
\alias{radarWidget}
\title{<Add Title>}
\usage{
radarWidget(df, legendCol = 1, firstColData = 2, showLegend = TRUE,
  legend.title = NULL, width = NULL, height = NULL)
}
\description{
<Add Description>
}

