% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radarWidget.R
\name{radarWidgetOutput}
\alias{radarWidgetOutput}
\title{Widget output function for use in Shiny}
\usage{
radarWidgetOutput(outputId, width = "100\%", height = "400px")
}
\description{
Widget output function for use in Shiny
}

