% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radarWidget.R
\name{renderRadarWidget}
\alias{renderRadarWidget}
\title{Widget render function for use in Shiny}
\usage{
renderRadarWidget(expr, env = parent.frame(), quoted = FALSE)
}
\description{
Widget render function for use in Shiny
}

