% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/radarWidget.R
\name{radarWidget}
\alias{radarWidget}
\title{radarWidget}
\usage{
radarWidget(df, legendCol = 1, firstColData = ifelse(is.null(legendCol), 1,
  2), legend.title = ifelse(is.null(legendCol), "Legend",
  colnames(df)[legendCol]), legend.text = df[, legendCol],
  showLegend = ifelse(is.null(legendCol), FALSE, TRUE),
  showTitle = showLegend, width = NULL, height = NULL)
}
\description{
radarWidget
}

